AskDIU/api/requirements.txt
pandas
python-dotenv
openai
runpod
pinecone
groq